src/daemon/migrator/requirements.txt
psycopg2-binary
pika
requests
